Fix call with ctypes, do not hardcode signature

+++ source.py
@@ -5,7 +5,7 @@ def ret_var() -> float:
     a: float = 42
     return a
 
-def ops() -> float:
+def ops() -> int:
     a: int = 2
     b: int = 4
     c: int = (a + b) * 2 - 3
